+++ store/driver/txn/unionstore_driver.go
@@ -22,71 +22,6 @@ import (
 	"github.com/pingcap/tidb/store/tikv/unionstore"
 )
 
-// tikvBatchGetter wraps unionstore.BufferBatchGetter as kv.BatchGetter.
-type tikvBatchGetter struct {
-	*unionstore.BufferBatchGetter
-}
-
-func (getter *tikvBatchGetter) BatchGet(ctx context.Context, keys []kv.Key) (map[string][]byte, error) {
-	return getter.BufferBatchGetter.BatchGet(ctx, toTiKVKeys(keys))
-}
-
-// kvBatchGetter wraps kv.BatchGetter as unionstore.BatchGetter.
-type kvBatchGetter struct {
-	kv.BatchGetter
-}
-
-func newUnionstoreBatchGetter(g kv.BatchGetter) unionstore.BatchGetter {
-	if g == nil {
-		return nil
-	}
-	return &kvBatchGetter{BatchGetter: g}
-}
-
-func (g *kvBatchGetter) BatchGet(ctx context.Context, keys [][]byte) (map[string][]byte, error) {
-	return g.BatchGetter.BatchGet(ctx, toKVKeys(keys))
-}
-
-// kvGetter wraps kv.Getter as unionstore.Getter
-type kvGetter struct {
-	kv.Getter
-}
-
-func newUnionstoreGetter(g kv.Getter) unionstore.Getter {
-	if g == nil {
-		return nil
-	}
-	return &kvGetter{Getter: g}
-}
-
-func (g *kvGetter) Get(k []byte) ([]byte, error) {
-	return g.Getter.Get(context.Background(), k)
-}
-
-type tikvMemBuffer struct {
-	kv.MemBuffer
-}
-
-func newUnionstoreBatchBufferGetter(m kv.MemBuffer) unionstore.BatchBufferGetter {
-	if m == nil {
-		return nil
-	}
-	return &tikvMemBuffer{MemBuffer: m}
-}
-
-func (g *tikvMemBuffer) Get(k []byte) ([]byte, error) {
-	return g.MemBuffer.Get(context.Background(), k)
-}
-
-// NewBufferBatchGetter creates a new BufferBatchGetter.
-func NewBufferBatchGetter(buffer kv.MemBuffer, middleCache kv.Getter, snapshot kv.BatchGetter) kv.BatchGetter {
-	tikvBuffer := newUnionstoreBatchBufferGetter(buffer)
-	tikvMiddleCache := newUnionstoreGetter(middleCache)
-	tikvSnapshot := newUnionstoreBatchGetter(snapshot)
-	getter := unionstore.NewBufferBatchGetter(tikvBuffer, tikvMiddleCache, tikvSnapshot)
-	return &tikvBatchGetter{getter}
-}
-
 // memBuffer wraps unionstore.MemDB as kv.MemBuffer.
 type memBuffer struct {
 	*unionstore.MemDB
