+++ store/tikv/unionstore/memdb_test.go
@@ -16,7 +16,6 @@
 package unionstore
 
 import (
-	"context"
 	"encoding/binary"
 	"fmt"
 	"testing"
@@ -826,31 +825,31 @@ func (s *testKVSuite) TestBufferLimit(c *C) {
 	c.Assert(err, NotNil)
 }
 
-func (s *testKVSuite) TestBufferBatchGetter(c *C) {
-	snap := &mockSnapshot{store: newMemDB()}
-	ka := []byte("a")
-	kb := []byte("b")
-	kc := []byte("c")
-	kd := []byte("d")
-	snap.store.Set(ka, ka)
-	snap.store.Set(kb, kb)
-	snap.store.Set(kc, kc)
-	snap.store.Set(kd, kd)
+// func (s *testKVSuite) TestBufferBatchGetter(c *C) {
+// 	snap := &mockSnapshot{store: newMemDB()}
+// 	ka := []byte("a")
+// 	kb := []byte("b")
+// 	kc := []byte("c")
+// 	kd := []byte("d")
+// 	snap.store.Set(ka, ka)
+// 	snap.store.Set(kb, kb)
+// 	snap.store.Set(kc, kc)
+// 	snap.store.Set(kd, kd)
 
-	// middle value is the same as snap
-	middle := newMemDB()
-	middle.Set(ka, []byte("a1"))
-	middle.Set(kc, []byte("c1"))
+// 	// middle value is the same as snap
+// 	middle := newMemDB()
+// 	middle.Set(ka, []byte("a1"))
+// 	middle.Set(kc, []byte("c1"))
 
-	buffer := newMemDB()
-	buffer.Set(ka, []byte("a2"))
-	buffer.Delete(kb)
+// 	buffer := newMemDB()
+// 	buffer.Set(ka, []byte("a2"))
+// 	buffer.Delete(kb)
 
-	batchGetter := NewBufferBatchGetter(buffer, middle, snap)
-	result, err := batchGetter.BatchGet(context.Background(), [][]byte{ka, kb, kc, kd})
-	c.Assert(err, IsNil)
-	c.Assert(len(result), Equals, 3)
-	c.Assert(string(result[string(ka)]), Equals, "a2")
-	c.Assert(string(result[string(kc)]), Equals, "c1")
-	c.Assert(string(result[string(kd)]), Equals, "d")
-}
+// 	batchGetter := NewBufferBatchGetter(buffer, middle, snap)
+// 	result, err := batchGetter.BatchGet(context.Background(), [][]byte{ka, kb, kc, kd})
+// 	c.Assert(err, IsNil)
+// 	c.Assert(len(result), Equals, 3)
+// 	c.Assert(string(result[string(ka)]), Equals, "a2")
+// 	c.Assert(string(result[string(kc)]), Equals, "c1")
+// 	c.Assert(string(result[string(kd)]), Equals, "d")
+// }
